requirements.txt
Authlib==0.13
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
google-api-python-client==1.7.11
google-auth==1.7.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
gspread==3.1.0
httplib2==0.14.0
idna==2.8
oauth2client==4.1.3
oauthlib==3.1.0
pprintpp==0.4.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
uritemplate==3.0.0
urllib3==1.25.7
